+++ front/src/app/components/forms/convention/ConventionEmailWarning.tsx
@@ -24,7 +24,7 @@ export const ConventionEmailWarning = ({
                 `/api${authRoutes.initiateLoginByOAuth.url}`,
                 {
                   provider: "ftConnect",
-                  redirectUri: `/${frontRoutes.conventionImmersion().href}`,
+                  redirectUri: frontRoutes.conventionImmersion().href,
                 },
               )}
             >
