+++ front/src/app/pages/convention/ConventionImmersionPage.tsx
@@ -190,7 +190,7 @@ const SharedConventionMessage = ({
             `/api${authRoutes.initiateLoginByOAuth.url}`,
             {
               provider: "ftConnect",
-              redirectUri: `/${frontRoutes.conventionImmersion().href}`,
+              redirectUri: frontRoutes.conventionImmersion().href,
             },
           )}
           onClick={() => {
